src/api/tmobile-api/tmobile/requirements.txt
boto3==1.13.16
tabula-py